+++ R/sleeper_userleagues.R
@@ -29,7 +29,7 @@ ff_userleagues.sleeper_conn <- function(conn = NULL, user_name = NULL, season = 
 
   if(is.null(season)) season <- .fn_choose_season()
 
-  df_leagues <- sleeper_getendpoint("user", user_id, "leagues/nfl", season) %>%
+  df_leagues <- sleeper_getendpoint(glue::glue("user/{user_id}/leagues/nfl/{season}")) %>%
     purrr::pluck("content") %>%
     purrr::map_dfr(`[`, c("name", "league_id")) %>%
     dplyr::rename(league_name = .data$name) %>%
@@ -62,7 +62,7 @@ sleeper_userleagues <- function(user_name, season = NULL){
 #' @noRd
 
 .sleeper_userteams <- function(league_id, user_id) {
-  df_teams <- sleeper_getendpoint("league", league_id, "users") %>%
+  df_teams <- sleeper_getendpoint(glue::glue("league/{league_id}/users")) %>%
     purrr::pluck("content") %>%
     tibble::tibble() %>%
     tidyr::hoist(1, "franchise_id" = "user_id", "display_name", "metadata") %>%